Split decoder to a seperate file
